[tvla] Move ttest fuctions into a separate file
This commit moves functions needed for computing t-tests from
tvla.py into a separate file.
The goal is to declutter tvla script to make it easier for
maintaining and debugging.
